+++ tasks/task1 - knn/Aditya_waghmare_KNN_algorithm.py
