+++ tasks/task3 - matrix_multiplication_gpu/task_3.txt
